requirements.txt
pillow==10.2.0
numpy==1.26.3
matplotlib==3.8.2
pytest==7.4.3
black==23.12.1
isort==5.13.2
flake8==6.1.0
pytest-cov>=4.1.0
